chore: removes cosmos dependant api end-points

+++ web/src/Web.App/Infrastructure/Apis/ApiRequests/PutCustomDataRequest.cs
@@ -43,7 +43,7 @@ public record PutCustomDataRequest
     public decimal? TotalPupils { get; set; }
     public decimal? PercentFreeSchoolMeals { get; set; }
     public decimal? PercentSpecialEducationNeeds { get; set; }
-    public int? TotalInternalFloorArea { get; set; }
+    public decimal? TotalInternalFloorArea { get; set; }
     public decimal? WorkforceFTE { get; set; }
     public decimal? TeachersFTE { get; set; }
     public decimal? SeniorLeadershipFTE { get; set; }
